[Refactor] Refactoring NaverWebtoon/
