+++ src/main/java/com/example/AOD/Webtoon/NaverWebtoon/config/NaverWebtoonConfig.java
@@ -3,15 +3,11 @@ package com.example.AOD.Webtoon.NaverWebtoon.config;
 import com.example.AOD.Webtoon.NaverWebtoon.crawler.NaverWebtoonCrawler;
 import com.example.AOD.Webtoon.NaverWebtoon.fetcher.WebtoonApiFetcher;
 import com.example.AOD.util.ChromeDriverProvider;
-import com.example.AOD.Webtoon.NaverWebtoon.util.NaverLoginHandler;
+import com.example.AOD.util.NaverLoginHandler;
 import org.springframework.context.annotation.Bean;
 
 //@Configuration
 public class NaverWebtoonConfig {
-    @Bean
-    public ChromeDriverProvider ChromeDriverProvider(){
-        return new ChromeDriverProvider();
-    }
 
     @Bean
     public WebtoonApiFetcher WebtoonApiFetcher(){
