+++ src/main/java/com/example/AOD/util/ChromeDriverProvider.java
@@ -26,7 +26,4 @@ public class ChromeDriverProvider {
         return new ChromeDriver(options);
     }
 
-
-
-
 }
